refactor(core): move DOM manipulation logic to its own file
This refactoring takes a step toward breaking down the
node_manipulation file into smaller, more focused files.
