+++ packages/core/src/render3/node_manipulation_i18n.ts
@@ -8,11 +8,12 @@
 
 import {assertDomNode, assertIndexInRange} from '../util/assert';
 
-import {TNode, TNodeFlags, TNodeType} from './interfaces/node';
+import {TNode, TNodeType} from './interfaces/node';
 import {Renderer} from './interfaces/renderer';
 import {RElement, RNode} from './interfaces/renderer_dom';
 import {LView} from './interfaces/view';
-import {getInsertInFrontOfRNodeWithNoI18n, nativeInsertBefore} from './node_manipulation';
+import {getInsertInFrontOfRNodeWithNoI18n} from './node_manipulation';
+import {nativeInsertBefore} from './dom_node_manipulation';
 import {unwrapRNode} from './util/view_utils';
 
 /**
