+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -521,7 +521,6 @@
   "nativeAppendChild",
   "nativeAppendOrInsertBefore",
   "nativeInsertBefore",
-  "nativeParentNode",
   "nextNgElementId",
   "ngOnChangesSetInput",
   "ngZoneInstanceId",
