+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -509,7 +509,6 @@
   "nativeAppendChild",
   "nativeAppendOrInsertBefore",
   "nativeInsertBefore",
-  "nativeParentNode",
   "nextBindingIndex",
   "nextNgElementId",
   "ngOnChangesSetInput",
