+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -597,7 +597,6 @@
   "nativeAppendChild",
   "nativeAppendOrInsertBefore",
   "nativeInsertBefore",
-  "nativeParentNode",
   "navigationCancelingError",
   "nextBindingIndex",
   "nextNgElementId",
