+++ packages/core/test/bundling/todo/bundle.golden_symbols.json
@@ -419,7 +419,6 @@
   "nativeAppendChild",
   "nativeAppendOrInsertBefore",
   "nativeInsertBefore",
-  "nativeParentNode",
   "nextBindingIndex",
   "nextNgElementId",
   "ngOnChangesSetInput",
